Update awsDeploy plugin to use new .zip package artifact
