+++ lib/plugins/aws/deploy/compile/functions/index.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const _ = require('lodash');
+const path = require('path');
 
 class AwsCompileFunctions {
   constructor(serverless, options) {
@@ -45,7 +46,7 @@ class AwsCompileFunctions {
         .S3Bucket =
         `${this.serverless.service.service}-${this.options.stage}-${this.options.region}`;
       newFunction.Properties.Code
-        .S3Key = ''; // will be replaced in a further step
+        .S3Key = this.serverless.service.package.artifact.split(path.sep).pop();
 
       if (!functionObject.handler) {
         const errorMessage = [
