+++ lib/plugins/aws/deploy/compile/functions/tests/index.js
@@ -40,7 +40,7 @@ describe('AwsCompileFunctions', () => {
         Properties: {
           Code: {
             S3Bucket: 'new-service-dev-us-east-1',
-            S3Key: '',
+            S3Key: 'artifact.zip',
           },
           FunctionName: 'new-service-dev-first',
           Handler: 'first.function.handler',
@@ -55,7 +55,7 @@ describe('AwsCompileFunctions', () => {
         Properties: {
           Code: {
             S3Bucket: 'new-service-dev-us-east-1',
-            S3Key: '',
+            S3Key: 'artifact.zip',
           },
           FunctionName: 'new-service-dev-second',
           Handler: 'second.function.handler',
@@ -78,6 +78,7 @@ describe('AwsCompileFunctions', () => {
     serverless.service.resources = { Resources: {} };
     awsCompileFunctions.serverless.service.functions = functionsObjectMock;
     awsCompileFunctions.serverless.service.service = 'new-service';
+    awsCompileFunctions.serverless.service.package.artifact = 'artifact.zip';
   });
 
   describe('#constructor()', () => {
