+++ lib/plugins/aws/deploy/index.js
@@ -4,6 +4,7 @@ const BbPromise = require('bluebird');
 const validate = require('./lib/validate');
 const initializeResources = require('./lib/initializeResources');
 const createStack = require('./lib/createStack');
+const uploadDeploymentPackage = require('./lib/uploadDeploymentPackage');
 const deployFunctions = require('./lib/deployFunctions');
 const updateStack = require('./lib/updateStack');
 
@@ -21,6 +22,7 @@ class AwsDeploy {
       validate,
       initializeResources,
       createStack,
+      uploadDeploymentPackage,
       deployFunctions,
       updateStack
     );
@@ -34,6 +36,7 @@ class AwsDeploy {
       'deploy:createProviderStacks': () => BbPromise.bind(this).then(this.createStack),
 
       'deploy:deploy': () => BbPromise.bind(this)
+        .then(this.uploadDeploymentPackage)
         .then(this.deployFunctions)
         .then(this.updateStack)
         .then(() => this.serverless.cli.log('Deployment successful!')),
