+++ lib/plugins/aws/deploy/lib/updateStack.js
@@ -2,25 +2,9 @@
 
 const BbPromise = require('bluebird');
 const async = require('async');
-const _ = require('lodash');
 
 module.exports = {
   update() {
-    this.serverless.cli.log('Adding function resources to CF...');
-    const serviceResources = this.serverless.service.resources;
-
-    _.forEach(serviceResources.Resources, (value, key) => {
-      if (_.find(this.deployedFunctions, { name: key })) {
-        const newValue = value;
-        newValue.Properties.Code.S3Key =
-          (_.find(this.deployedFunctions, { name: key }).zipFileKey);
-
-        _.merge(serviceResources.Resources, { [key]: newValue });
-      }
-    });
-
-    this.serverless.service.resources = serviceResources;
-
     this.serverless.cli.log('Updating Stack...');
     const stackName = `${this.serverless.service.service}-${this.options.stage}`;
     const params = {
