+++ lib/plugins/aws/deploy/tests/all.js
@@ -3,6 +3,7 @@
 require('./validate');
 require('./initializeResources');
 require('./createStack');
+require('./uploadDeploymentPackage');
 require('./deployFunctions');
 require('./updateStack');
 require('./index');
