+++ lib/plugins/aws/deploy/tests/index.js
@@ -52,15 +52,19 @@ describe('AwsDeploy', () => {
     });
 
     it('should run "deploy:deploy" promise chain in order', () => {
+      const uploadDeploymentPackage = sinon
+        .stub(awsDeploy, 'uploadDeploymentPackage').returns(BbPromise.resolve());
       const deployFunctionsStub = sinon
         .stub(awsDeploy, 'deployFunctions').returns(BbPromise.resolve());
       const updateStackStub = sinon
         .stub(awsDeploy, 'updateStack').returns(BbPromise.resolve());
 
       return awsDeploy.hooks['deploy:deploy']().then(() => {
-        expect(deployFunctionsStub.calledOnce).to.be.equal(true);
+        expect(uploadDeploymentPackage.calledOnce).to.be.equal(true);
+        expect(deployFunctionsStub.calledAfter(uploadDeploymentPackage)).to.be.equal(true);
         expect(updateStackStub.calledAfter(deployFunctionsStub)).to.be.equal(true);
 
+        awsDeploy.uploadDeploymentPackage.restore();
         awsDeploy.deployFunctions.restore();
         awsDeploy.updateStack.restore();
       });
