+++ lib/plugins/aws/deploy/tests/updateStack.js
@@ -35,7 +35,7 @@ describe('updateStack', () => {
         Properties: {
           Code: {
             S3Bucket: 'new-service-dev-us-east-1',
-            S3Key: '',
+            S3Key: 'zip-file.zip',
           },
           FunctionName: 'new-service-first',
           Handler: 'first.function.handler',
@@ -74,17 +74,6 @@ describe('updateStack', () => {
       updateStackStub = sinon
         .stub(awsDeploy.sdk, 'request').returns(BbPromise.resolve());
     });
-
-    it('should add the S3Key to the compiled function resources before updating the stack',
-      () => awsDeploy.update().then(() => {
-        expect(updateStackStub.calledOnce).to.be.equal(true);
-        expect(updateStackStub.calledWith(awsDeploy.options.stage, awsDeploy.options.region));
-        expect(awsDeploy.serverless.service.resources.Resources
-          .first.Properties.Code.S3Key).to.equal('zipFileOfFirstFunction');
-
-        awsDeploy.sdk.request.restore();
-      })
-    );
 
     it('should update the stack', () => awsDeploy.update()
       .then(() => {
